Create 109. Convert Sorted List to Binary Search Tree.py
Given the head of a singly linked list where elements are sorted in ascending order, convert it to a 
height-balanced
 binary search tree.

 

Example 1:


Input: head = [-10,-3,0,5,9]
Output: [0,-3,9,-10,null,5]
Explanation: One possible answer is [0,-3,9,-10,null,5], which represents the shown height balanced BST.
Example 2:

Input: head = []
Output: []
To solve this problem, we can use the same approach as the binary search. We can first find the middle node of the linked list and make it the root of the binary search tree. Then we can recursively do the same for the left and right halves of the linked list to construct the left and right subtrees of the root.

To find the middle node of the linked list, we can use the two-pointer technique. We can have two pointers, one slow and one fast, that move through the linked list at different speeds. When the fast pointer reaches the end of the list, the slow pointer will be pointing to the middle node.